+++ src/engine/entity/EntityManager.h
@@ -15,5 +15,5 @@ class EntityManager {
     void addEntity(Entity* e);
     void markForRemoval(Entity* e);
 
-    void update(uint32_t);
+    void update(double);
 };
